requirements.txt
yt-dlp
openai-whisper
ffmpeg-python
requests
python-dotenv
certifi
react-router-dom
@google/generative-ai
dotenv
axios